adapt to new plugin structure
