+++ localstack/services/dynamodb/resource_providers/aws_dynamodb_table.py
@@ -2,11 +2,10 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Optional, Type, TypedDict
+from typing import Optional, TypedDict
 
 import localstack.services.cloudformation.provider_utils as util
 from localstack.services.cloudformation.resource_provider import (
-    CloudFormationResourceProviderPlugin,
     OperationStatus,
     ProgressEvent,
     ResourceProvider,
@@ -413,13 +412,3 @@ class DynamoDBTableProvider(ResourceProvider[DynamoDBTableProperties]):
         if args:
             args["TableName"] = properties["TableName"]
         return args
-
-
-class DynamoDBTableProviderPlugin(CloudFormationResourceProviderPlugin):
-    name = "AWS::DynamoDB::Table"
-
-    def __init__(self):
-        self.factory: Optional[Type[ResourceProvider]] = None
-
-    def load(self):
-        self.factory = DynamoDBTableProvider
