+++ localstack/services/dynamodb/resource_providers/aws_dynamodb_table_plugin.py
@@ -1,0 +1,20 @@
+from typing import Optional, Type
+
+from localstack.services.cloudformation.resource_provider import (
+    CloudFormationResourceProviderPlugin,
+    ResourceProvider,
+)
+
+
+class DynamoDBTableProviderPlugin(CloudFormationResourceProviderPlugin):
+    name = "AWS::DynamoDB::Table"
+
+    def __init__(self):
+        self.factory: Optional[Type[ResourceProvider]] = None
+
+    def load(self):
+        from localstack.services.dynamodb.resource_providers.aws_dynamodb_table import (
+            DynamoDBTableProvider,
+        )
+
+        self.factory = DynamoDBTableProvider
